Fixed compiler warnings
